blog list - 4.15-4.23

+++ controllers/blogs.js
@@ -1,8 +1,13 @@
+// const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+// const User = require('../models/user')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({}).populate('user', { username: 1, name: 1 })
+
     response.json(blogs)
 })
 
@@ -16,8 +21,11 @@ blogsRouter.get('/:id', async (request, response) => {
     }
 })
 
-blogsRouter.post('/', async (request, response) => {
+
+//Take note of userExractor and the import above for a specific route - Chris
+blogsRouter.post('/', userExtractor, async (request, response) => {
     const body = request.body
+    const user = request.user
 
     if (!body.title || !body.url) {
         response.status(400).end()
@@ -27,16 +35,28 @@ blogsRouter.post('/', async (request, response) => {
             author: body.author,
             url: body.url,
             likes: body.likes || 0,
+            user: user.id
         })
 
         const savedBlog = await blog.save()
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
+
         response.status(201).json(savedBlog)
     }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+
+    const userid = request.user.id
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === userid.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+        return response.status(401).json({ error: 'User does not match the original user that posted the blog' })
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
